"Refactor project structure and packages
- Moved model, service, exceptions, and util packages under app/
- Renamed database package to repository for clarity
- Moved DatabaseManager and DatabaseSetup to app.repository
- Renamed AccountDBHelper to AccountRepository
- Updated package declarations and imports accordingly
- Updated README with new project structure and run instructions"
