+++ config.properties
@@ -2,5 +2,5 @@
 # File used to store serialized accounts
 account.file.path=data/accounts.txt
 
-# SQLite database file path
+# SQLite app.database file path
 accountDatabase.file.path=data/accountsDB.db
