+++ src/app/exceptions/GoBackToMainMenuException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package app.exceptions;
 
 public class GoBackToMainMenuException extends Exception {
     // This class will be used whenever the user enters 'back' text
