+++ src/app/repository/DatabaseSetup.java
@@ -1,4 +1,6 @@
-package database;
+package app.repository;
+import app.repository.DatabaseManager;
+
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
